test: move ng_config utils to seperate module
